Move lobby.server.db package to lobby project
