+++ build.gradle
@@ -55,6 +55,7 @@ subprojects {
 
         errorprone 'com.google.errorprone:error_prone_core:2.3.1'
 
+        testCompile 'nl.jqno.equalsverifier:equalsverifier:2.4.5'
         testCompile "org.hamcrest:java-hamcrest:$hamcrestVersion"
         testCompile "org.mockito:mockito-core:$mockitoVersion"
         testCompile "org.mockito:mockito-junit-jupiter:$mockitoVersion"
